% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fn.R
\name{test_fn}
\alias{test_fn}
\title{Test function}
\usage{
test_fn(x)
}
\arguments{
\item{x}{A number}
}
\value{
x incremented by 1
}
\description{
This function adds one to a number.
}
\examples{
num_plus_one <- test_fn(1)

}
\author{
Andrew Shockley
}
